subgroup panels into categories
