+++ components/panels/__init__.py
@@ -1,0 +1,5 @@
+from .correcting_agent import correcting_agent_panel
+from .genetics import genetics_panel, gene_panel
+from .rag_agent import rag_agent_panel
+from .kg import kg_panel
+from .project import summary_panel, tasks_panel
